% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer-package.R
\docType{package}
\name{cer-package}
\alias{cer}
\alias{cer-package}
\title{cer: Australian Clean Energy Regulator Data}
\description{
Tidy R access to data published by the Australian Clean Energy
Regulator on carbon credits, the Safeguard Mechanism, renewable
energy certificates, and greenhouse gas reporting.
}
\section{Main function families}{

\itemize{
\item \code{\link[=cer_accu_projects]{cer_accu_projects()}}, \code{\link[=cer_accu_issuances]{cer_accu_issuances()}},
\code{\link[=cer_accu_contracts]{cer_accu_contracts()}}, \code{\link[=cer_accu_methods]{cer_accu_methods()}},
\code{\link[=cer_accu_relinquishments]{cer_accu_relinquishments()}}: Australian Carbon Credit Unit
('ACCU') Scheme.
\item \code{\link[=cer_safeguard_facilities]{cer_safeguard_facilities()}}: Safeguard Mechanism covered
emissions, baselines, and SMC issuance.
\item \code{\link[=cer_nger_corporate]{cer_nger_corporate()}}, \code{\link[=cer_nger_electricity]{cer_nger_electricity()}}: National
Greenhouse and Energy Reporting ('NGER') scheme.
\item \code{\link[=cer_lgc_power_stations]{cer_lgc_power_stations()}}: Large-scale Renewable Energy Target
('LRET') accredited power stations.
\item \code{\link[=cer_sres_installations]{cer_sres_installations()}}: Small-scale Renewable Energy Scheme
('SRES') postcode installation data.
\item \code{\link[=cer_qcmr]{cer_qcmr()}}: Quarterly Carbon Market Report data workbook.
\item \code{\link[=cer_cache_info]{cer_cache_info()}}, \code{\link[=cer_clear_cache]{cer_clear_cache()}}: cache management.
}
}

\section{Data source}{

All data is published by the Clean Energy Regulator at
\url{https://cer.gov.au/markets/reports-and-data} under a Creative
Commons Attribution 4.0 International licence.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://charlescoverdale.github.io/cer/}
  \item \url{https://github.com/charlescoverdale/cer}
  \item Report bugs at \url{https://github.com/charlescoverdale/cer/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\concept{ACCU}
\concept{Australian climate policy}
\concept{NGER}
\concept{Safeguard Mechanism}
\concept{carbon markets}
\keyword{internal}
