% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{cer_manifest_clear}
\alias{cer_manifest_clear}
\title{Clear the session manifest}
\usage{
cer_manifest_clear()
}
\value{
Invisibly \code{NULL}.
}
\description{
Clear the session manifest
}
\examples{
cer_manifest_clear()
}
\seealso{
Other reproducibility: 
\code{\link{cer_cite}()},
\code{\link{cer_deposit_zenodo}()},
\code{\link{cer_legislative_instrument}()},
\code{\link{cer_manifest}()},
\code{\link{cer_manifest_write}()},
\code{\link{cer_sha256}()},
\code{\link{cer_snapshot}()}
}
\concept{reproducibility}
