% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safeguard_reform.R
\name{cer_safeguard_industry_baselines}
\alias{cer_safeguard_industry_baselines}
\title{Bundled Safeguard industry-average default baselines}
\usage{
cer_safeguard_industry_baselines()
}
\value{
A \code{cer_tbl} of industry default baselines.
}
\description{
Bundled Safeguard industry-average default baselines
}
\examples{
cer_safeguard_industry_baselines()
}
\seealso{
Other safeguard: 
\code{\link{cer_safeguard_baseline_trajectory}()},
\code{\link{cer_safeguard_facilities}()},
\code{\link{cer_safeguard_teba_facilities}()}
}
\concept{safeguard}
